backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.0
pydantic==2.9.0
claude-agent-sdk>=0.1.19
pytest==8.0.0
pytest-asyncio==0.23.0
